requirements.txt
flask
SQLAlchemy
requests
APScheduler
pillow
Flask-Login
Flask-RESTful
python-dotenv
sqlalchemy_serializer
aiogram==2.25.1